requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.4
django-cors-headers==4.6.0
django-decouple==2.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
openapi-codec==1.3.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
setuptools==75.6.0
simplejson==3.19.3
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.2.3
